refactor: #252 - DependencyKey, FollowUseCase 분리
