+++ Spoony-iOS/Spoony-iOS/Resource/Extension/DependencyValues+.swift
@@ -42,4 +42,9 @@ extension DependencyValues {
         get { self[ImageLoadServiceKey.self] }
         set { self[ImageLoadServiceKey.self] = newValue }
     }
+    
+    var followUseCase: FollowUseCase {
+        get { self[FollowUseCaseKey.self] }
+        set { self[FollowUseCaseKey.self] = newValue }
+    }
 }
